+++ trixnity-messenger-compose-view/src/desktopMain/kotlin/de/connect2x/messenger/desktop/messengerApp.kt
@@ -32,7 +32,7 @@ import de.connect2x.messenger.compose.view.IsFocused
 import de.connect2x.messenger.compose.view.Platform
 import de.connect2x.messenger.compose.view.PlatformType
 import de.connect2x.messenger.compose.view.notifications.Notifications
-import de.connect2x.messenger.compose.view.profiles.Profiles
+import de.connect2x.messenger.compose.view.profiles.IntroductionOrProfile
 import de.connect2x.messenger.compose.view.profiles.ShowProfileCreation
 import de.connect2x.messenger.compose.view.profiles.WithProfileSelection
 import de.connect2x.messenger.compose.view.theme.IsFocusHighlighting
@@ -172,7 +172,7 @@ fun CoroutineScope.messengerApp(
                         EscapeKeyPressed provides escapeKeyPressed,
                     ) {
                         MessengerTheme {
-                            Profiles()
+                            IntroductionOrProfile()
                         }
                     }
                 }